completed client server communication

+++ include/message_handler.h
@@ -1,15 +1,15 @@
 #pragma once
 #include "message.h"
 
-Message generate_upload_file_message(const uint8_t filename, const uint8_t port);
+Message generate_upload_file_message(const uint32_t filename, const uint32_t port);
 
-Message generate_delete_file_message(const uint8_t filename, const uint8_t port);
+Message generate_delete_file_message(const uint32_t filename, const uint32_t port);
 
-Message generate_download_file_message(const uint8_t filename);
+Message generate_download_file_message(const uint32_t filename);
 
 Message generate_handshake_message();
 
-Message generate_file_request_message(const uint8_t filename);
+Message generate_file_request_message(const uint32_t filename);
 
 Message generate_bitfields_message(const std::vector<uint8_t>& bitfield);
